notebooks/pyproject.toml
[project]
name = "notebooks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "azure-identity>=1.25.1",
    "azure-search-documents>=11.6.0",
    "python-dotenv>=1.2.1",
]
